update routes & layout
